requirements.txt
feedparser
requests 
beautifulsoup4 
fastapi 
uvicorn 
pytz
sqlalchemy 
pydantic
newspaper3k 
